feat: release_date 필드 수정

+++ server/src/main/java/com/cinemaprincess/movie/dto/MovieDetailResponseDto.java
@@ -2,6 +2,8 @@ package com.cinemaprincess.movie.dto;
 
 import com.cinemaprincess.genre.Genre;
 import com.cinemaprincess.genre.GenreDto;
+import com.cinemaprincess.review.dto.ReviewResponseDto;
+import com.cinemaprincess.review.entity.Review;
 import com.cinemaprincess.watch_provider.WatchProviderDto;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,11 +13,11 @@ import java.util.List;
 @Getter
 @Builder
 public class MovieDetailResponseDto {
-    private long movieId;
+    private Long movieId;
     private String backdropPath;
     private String posterPath;
     private String title;
-    private double voteAverage;
+    private float voteAverage;
     private String releaseDate;
     private String overview;
     private Integer runtime;
@@ -25,4 +27,5 @@ public class MovieDetailResponseDto {
     private String videoPath;
     private List<GenreDto.Response> genres;
     private List<WatchProviderDto.Response> watchProviders;
+//    private List<ReviewResponseDto> reviews;
 }
